# Updating to use pure go xz lib