--- a/.travis.yml
+++ b/.travis.yml
@@ -1,11 +1,7 @@
 language: go
-go: 
+go:
  - 1.6
- - 1.7 
+ - 1.7
 
-before_install:
- - sudo apt-get update -qq
- - sudo apt-get install -qq liblzma-dev
- 
 script:
  - go test -v ./...