--- a/xz_cgo_reader.go
+++ b/xz_cgo_reader.go
@@ -1,0 +1,21 @@
+// +build cgo
+
+package zim
+
+import (
+	"io"
+
+	xz "github.com/remyoudompheng/go-liblzma"
+)
+
+type XZReader struct {
+	*xz.Decompressor
+}
+
+func NewXZReader(r io.Reader) (*XZReader, error) {
+	dec, err := xz.NewReader(r)
+	if err != nil {
+		return nil, err
+	}
+	return &XZReader{dec}, nil
+}