Hi.  Thanks for the great lib.  I switched to a pure go xz lib for easier cross-compillation.  It passes tests and works great for me.  Feel free to merge or discard if you're happier with the existing lib.

Thanks!
Jason