@akhenakh 2018-02-02
Hey thanks for the PR, unfortunately pure xz is terribly slow, look at this previous PR: https://github.com/akhenakh/gozim/pull/32

Can you please check the performance and see if it's better now ?