@jasonmoo 2018-02-02
Hi Cool totally missed that old PR.  You are correct the pure go implementation is ~2.5x slower.  

I need the cross-compilation ability so I will be keeping the change, but I understand if you want to prioritize performance in your lib.  It's a shame the perf difference is so significant.

I added a benchmark that clears the LRU and tested both libs.  The memory increase is harder to measure due to cgo obfuscating what the underlying c code is allocing.  
Benchcmp for 10 runs on each lib.  
```
jason@mba ~/go/src/github.com/jasonmoo/gozim master: benchcmp -best cgo.txt go.txt 
benchmark                   old ns/op     new ns/op     delta
BenchmarkArticleBytes-4     29053948      78114929      +168.86%

benchmark                   old MB/s     new MB/s     speedup
BenchmarkArticleBytes-4     2.72         1.01         0.37x

benchmark                   old allocs     new allocs     delta
BenchmarkArticleBytes-4     126            215620         +171026.98%

benchmark                   old bytes     new bytes     delta
BenchmarkArticleBytes-4     6439808       75718960      +1075.80%
```