@akhenakh 2018-02-02
2.5x diff would be ok, but I know some ppl, myself included,are using it on giant gozim like wikipedia, where the performance hit would be too high.

Is the API for pure xz exactly the same, maybe we could incorporate your changes with a build flags ?