@jasonmoo 2018-02-07
I added the change behind a build tag.  This defaults to the pure go lib when cross-compiling but the cgo lib when compiling natively.